+++ src/config.py
@@ -28,7 +28,8 @@ from pathlib import Path
 rocprof_compute_home = Path(__file__).resolve().parent
 PROJECT_NAME = "rocprofiler-compute"
 
-HIDDEN_COLUMNS = ["Tips", "coll_level"]
+HIDDEN_COLUMNS = ["coll_level"]
+HIDDEN_COLUMNS_CLI = ["Description", "coll_level"]
 HIDDEN_SECTIONS = [400, 1900, 2000]
 
 TIME_UNITS = {"s": 10**9, "ms": 10**6, "us": 10**3, "ns": 1}